Move 'configureJvmTestCommon' to 'Testing.kt'
